+++ example/src/pages/button-group/ButtonGroup.svelte
@@ -8,9 +8,15 @@
 	</div>
 
 	<div class="demo-content">
+		<p>
+			Button with
+			<code>'toggle'</code>
+			property set
+		</p>
 		<ButtonGroup>
-			<ToggleButton
+			<Button
 				active={justify.left}
+				toggle
 				{...props}
 				color="accent"
 				on:change={(e) => {
@@ -18,9 +24,10 @@
 				}}
 			>
 				<Icon path={alignLeft} />
-			</ToggleButton>
-			<ToggleButton
+			</Button>
+			<Button
 				active={justify.center}
+				toggle
 				{...props}
 				color="accent"
 				on:change={(e) => {
@@ -28,9 +35,10 @@
 				}}
 			>
 				<Icon path={alignCenter} />
-			</ToggleButton>
-			<ToggleButton
+			</Button>
+			<Button
 				active={justify.right}
+				toggle
 				{...props}
 				color="accent"
 				on:change={(e) => {
@@ -38,9 +46,10 @@
 				}}
 			>
 				<Icon path={alignRight} />
-			</ToggleButton>
-			<ToggleButton
+			</Button>
+			<Button
 				active={justify.justify}
+				toggle
 				{...props}
 				color="accent"
 				on:change={(e) => {
@@ -48,14 +57,14 @@
 				}}
 			>
 				<Icon path={alignJustify} />
-			</ToggleButton>
+			</Button>
 
-			<ToggleButton bind:active={isbold} {...props} color="accent">
+			<Button bind:active={isbold} toggle {...props} color="accent">
 				<Icon path={bold} />
-			</ToggleButton>
-			<ToggleButton bind:active={isitalic} {...props} color="accent">
+			</Button>
+			<Button bind:active={isitalic} toggle {...props} color="accent">
 				<Icon path={italic} />
-			</ToggleButton>
+			</Button>
 		</ButtonGroup>
 	</div>
 
@@ -97,7 +106,7 @@
 </Description>
 
 <script>
-	import { ButtonGroup, Button, ToggleButton, Checkbox, Icon, Radio } from '@vikignt/svelte-ui';
+	import { ButtonGroup, Button, Checkbox, Icon, Radio } from '@vikignt/svelte-ui';
 	import {
 		alignCenter,
 		alignJustify,
@@ -143,12 +152,8 @@
 	}
 
 	function onjustify(param, value) {
-		if (!value) {
-			justify[param] = value;
-			return;
-		}
 		Object.keys(justify).map((key) => {
-			justify[key] = key === param ? true : false;
+			justify[key] = key === param ? value : false;
 		});
 	}
 </script>